Fixed misc. warnings picked up by the static analyser and compile errors from removed old ISL backend.
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@2741 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ networking/packet.cpp
@@ -89,7 +89,7 @@ void Packet::write(const void* src, size_t numbytes)
 	const size_t writeindex = data.size();
 	data.resize(writeindex + numbytes);
 
-	for(int i=0; i<numbytes; ++i)
+	for(size_t i=0; i<numbytes; ++i)
 	{
 		data[writeindex + i] = *((char*)src + i);
 	}
